feat: split dividend from equity, new DividendFacet
