+++ packages/ats/contracts/contracts/domain/asset/equity/IEquityStorageWrapper.sol
@@ -2,16 +2,6 @@
 pragma solidity >=0.8.0 <0.9.0;
 
 interface IEquityStorageWrapper {
-    event DividendSet(
-        bytes32 corporateActionId,
-        uint256 dividendId,
-        address indexed operator,
-        uint256 indexed recordDate,
-        uint256 indexed executionDate,
-        uint256 amount,
-        uint8 amountDecimals
-    );
-
     event ScheduledBalanceAdjustmentSet(
         bytes32 corporateActionId,
         uint256 balanceAdjustmentId,
@@ -22,21 +12,11 @@ interface IEquityStorageWrapper {
     );
 
     /**
-     * @notice Emitted when a dividend is cancelled.
-     * @param dividendId Identifier of the cancelled dividend.
-     * @param operator Address that performed the cancellation.
-     */
-    event DividendCancelled(uint256 dividendId, address indexed operator);
-
-    /**
      * @notice Emitted when a scheduled balance adjustment is cancelled.
      * @param balanceAdjustmentId Identifier of the cancelled scheduled balance adjustment.
      * @param operator Address that performed the cancellation.
      */
     event ScheduledBalanceAdjustmentCancelled(uint256 balanceAdjustmentId, address indexed operator);
-
-    error DividendCreationFailed();
     error BalanceAdjustmentCreationFailed();
-    error DividendAlreadyExecuted(bytes32 corporateActionId, uint256 dividendId);
     error BalanceAdjustmentAlreadyExecuted(bytes32 corporateActionId, uint256 balanceAdjustmentId);
 }
