+++ packages/ats/contracts/contracts/facets/layer_2/dividend/standard/DividendFacet.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+pragma solidity >=0.8.0 <0.9.0;
+
+import { DividendFacetBase } from "../DividendFacetBase.sol";
+import { _DIVIDEND_RESOLVER_KEY } from "../../../../constants/resolverKeys.sol";
+import { Common } from "../../../../domain/Common.sol";
+
+contract DividendFacet is DividendFacetBase, Common {
+    function getStaticResolverKey() external pure override returns (bytes32 staticResolverKey_) {
+        staticResolverKey_ = _DIVIDEND_RESOLVER_KEY;
+    }
+}
