+++ packages/ats/contracts/contracts/facets/layer_3/equityUSA/EquityUSAFacet.sol
@@ -15,15 +15,10 @@ contract EquityUSAFacet is EquityUSA, IStaticFunctionSelectors {
 
     function getStaticFunctionSelectors() external pure override returns (bytes4[] memory staticFunctionSelectors_) {
         uint256 selectorIndex;
-        staticFunctionSelectors_ = new bytes4[](17);
+
+        staticFunctionSelectors_ = new bytes4[](9);
         staticFunctionSelectors_[selectorIndex++] = this._initialize_equityUSA.selector;
         staticFunctionSelectors_[selectorIndex++] = this.getEquityDetails.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.setDividend.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.cancelDividend.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getDividend.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getDividendFor.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getDividendAmountFor.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getDividendsCount.selector;
         staticFunctionSelectors_[selectorIndex++] = this.setScheduledBalanceAdjustment.selector;
         staticFunctionSelectors_[selectorIndex++] = this.cancelScheduledBalanceAdjustment.selector;
         staticFunctionSelectors_[selectorIndex++] = this.getScheduledBalanceAdjustment.selector;
@@ -31,8 +26,6 @@ contract EquityUSAFacet is EquityUSA, IStaticFunctionSelectors {
         staticFunctionSelectors_[selectorIndex++] = this.getSecurityRegulationData.selector;
         staticFunctionSelectors_[selectorIndex++] = this.getSecurityHolders.selector;
         staticFunctionSelectors_[selectorIndex++] = this.getTotalSecurityHolders.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getDividendHolders.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getTotalDividendHolders.selector;
     }
 
     function getStaticInterfaceIds() external pure override returns (bytes4[] memory staticInterfaceIds_) {
