+++ packages/ats/contracts/contracts/test/testTimeTravel/facetsTimeTravel/dividend/DividendFacetTimeTravel.sol
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: Apache-2.0
+pragma solidity >=0.8.0 <0.9.0;
+
+import { DividendFacet } from "../../../../facets/layer_2/dividend/standard/DividendFacet.sol";
+import { TimeTravelStorageWrapper } from "../../timeTravel/TimeTravelStorageWrapper.sol";
+import { LocalContext } from "../../../../infrastructure/utils/LocalContext.sol";
+
+contract DividendFacetTimeTravel is DividendFacet, TimeTravelStorageWrapper {
+    function _blockTimestamp() internal view override(LocalContext, TimeTravelStorageWrapper) returns (uint256) {
+        return TimeTravelStorageWrapper._blockTimestamp();
+    }
+
+    function _blockNumber() internal view override(LocalContext, TimeTravelStorageWrapper) returns (uint256) {
+        return TimeTravelStorageWrapper._blockNumber();
+    }
+}
