+++ packages/ats/contracts/test/contracts/integration/layer_1/ERC1400/ERC1410/erc1410.test.ts
@@ -9,6 +9,7 @@ import {
   ClearingActionsFacet,
   ControlListFacet,
   DiamondFacet,
+  DividendFacet,
   type Equity,
   ERC1594Facet,
   ERC1644Facet,
@@ -112,6 +113,7 @@ describe("ERC1410 Tests", () => {
   let snapshotsFacet: SnapshotsFacet;
   let timeTravelFacet: TimeTravelFacet;
   let diamondCutFacet: DiamondFacet;
+  let dividendFacet: DividendFacet;
 
   async function setPreBalanceAdjustment(singlePartition?: boolean) {
     await grantRolesToAccounts();
@@ -365,7 +367,7 @@ describe("ERC1410 Tests", () => {
     clearingActionsFacet = await ethers.getContractAt("ClearingActionsFacet", diamond.target, signer_A);
     snapshotsFacet = await ethers.getContractAt("SnapshotsFacet", diamond.target);
     diamondCutFacet = await ethers.getContractAt("DiamondFacet", diamond.target);
-
+    dividendFacet = await ethers.getContractAt("DividendFacet", diamond.target, signer_A);
     capFacet = await ethers.getContractAt("Cap", diamond.target);
 
     timeTravelFacet = await ethers.getContractAt("TimeTravelFacet", diamond.target);
@@ -1178,8 +1180,8 @@ describe("ERC1410 Tests", () => {
         amount: 1,
         amountDecimals: 0,
       };
-      await equityFacet.connect(signer_C).setDividend(dividendData_1);
-      await equityFacet.connect(signer_C).setDividend(dividendData);
+      await dividendFacet.connect(signer_C).setDividend(dividendData_1);
+      await dividendFacet.connect(signer_C).setDividend(dividendData);
 
       //  transfer
       const canTransfer = await erc1410Facet
@@ -1206,13 +1208,13 @@ describe("ERC1410 Tests", () => {
       expect(balanceOf_E).to.equal(balanceOf_E_Original - amount);
       const balanceOf_D = await erc1410Facet.balanceOf(signer_D.address);
       expect(balanceOf_D).to.equal(2 * amount);
-      let dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      let dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      let dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      let dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(0);
       expect(dividend.snapshotId).to.equal(0);
-      let dividend_1_For_C = await equityFacet.getDividendFor(1, signer_C.address);
-      let dividend_1_For_E = await equityFacet.getDividendFor(1, signer_E.address);
-      let dividend_1_For_D = await equityFacet.getDividendFor(1, signer_D.address);
+      let dividend_1_For_C = await dividendFacet.getDividendFor(1, signer_C.address);
+      let dividend_1_For_E = await dividendFacet.getDividendFor(1, signer_E.address);
+      let dividend_1_For_D = await dividendFacet.getDividendFor(1, signer_D.address);
       expect(dividend_1_For_C.tokenBalance).to.equal(0);
       expect(dividend_1_For_E.tokenBalance).to.equal(0);
       expect(dividend_1_For_D.tokenBalance).to.equal(0);
@@ -1225,12 +1227,12 @@ describe("ERC1410 Tests", () => {
       // AFTER FIRST SCHEDULED SNAPSHOTS ------------------------------------------------------------------
       await timeTravelFacet.changeSystemTimestamp(dividendsRecordDateInSeconds_1 + 1);
 
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(0);
 
-      dividend_1_For_C = await equityFacet.getDividendFor(1, signer_C.address);
-      dividend_1_For_E = await equityFacet.getDividendFor(1, signer_E.address);
-      dividend_1_For_D = await equityFacet.getDividendFor(1, signer_D.address);
+      dividend_1_For_C = await dividendFacet.getDividendFor(1, signer_C.address);
+      dividend_1_For_E = await dividendFacet.getDividendFor(1, signer_E.address);
+      dividend_1_For_D = await dividendFacet.getDividendFor(1, signer_D.address);
 
       expect(dividend_1_For_C.tokenBalance).to.equal(balanceOf_C);
       expect(dividend_1_For_E.tokenBalance).to.equal(balanceOf_E);
@@ -1247,13 +1249,13 @@ describe("ERC1410 Tests", () => {
         .to.emit(snapshotsFacet, "SnapshotTriggered")
         .withArgs(1, ethers.toBeHex(1, 32));
       // check that scheduled snapshots was triggered
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(1);
       expect(dividend.snapshotId).to.equal(0);
-      dividend_1_For_C = await equityFacet.getDividendFor(1, signer_C.address);
-      dividend_1_For_E = await equityFacet.getDividendFor(1, signer_E.address);
-      dividend_1_For_D = await equityFacet.getDividendFor(1, signer_D.address);
+      dividend_1_For_C = await dividendFacet.getDividendFor(1, signer_C.address);
+      dividend_1_For_E = await dividendFacet.getDividendFor(1, signer_E.address);
+      dividend_1_For_D = await dividendFacet.getDividendFor(1, signer_D.address);
 
       expect(dividend_1_For_C.tokenBalance).to.equal(balanceOf_C);
       expect(dividend_1_For_E.tokenBalance).to.equal(balanceOf_E);
@@ -1274,8 +1276,8 @@ describe("ERC1410 Tests", () => {
         .withArgs(2, ethers.toBeHex(2, 32));
 
       // check that scheduled snapshots was triggered
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(1);
       expect(dividend.snapshotId).to.equal(2);
     });
@@ -1329,8 +1331,8 @@ describe("ERC1410 Tests", () => {
         amount: 1,
         amountDecimals: 0,
       };
-      await equityFacet.connect(signer_C).setDividend(dividendData_1);
-      await equityFacet.connect(signer_C).setDividend(dividendData);
+      await dividendFacet.connect(signer_C).setDividend(dividendData_1);
+      await dividendFacet.connect(signer_C).setDividend(dividendData);
 
       //  transfer
       await expect(
@@ -1356,8 +1358,8 @@ describe("ERC1410 Tests", () => {
       const totalSupplyByPartition = await erc1410Facet.totalSupplyByPartition(_PARTITION_ID_1);
       expect(totalSupply).to.equal(BigInt(balanceOf_C_Original) + BigInt(balanceOf_E_Original) + balanceOf_D);
       expect(totalSupplyByPartition.toString()).to.equal(totalSupply.toString());
-      let dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      let dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      let dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      let dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(0);
       expect(dividend.snapshotId).to.equal(0);
 
@@ -1385,8 +1387,8 @@ describe("ERC1410 Tests", () => {
         .withArgs(1, ethers.toBeHex(1, 32));
 
       // check that scheduled snapshots was triggered
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(1);
       expect(dividend.snapshotId).to.equal(0);
     });
@@ -1411,8 +1413,8 @@ describe("ERC1410 Tests", () => {
         amount: 1,
         amountDecimals: 0,
       };
-      await equityFacet.connect(signer_C).setDividend(dividendData_1);
-      await equityFacet.connect(signer_C).setDividend(dividendData);
+      await dividendFacet.connect(signer_C).setDividend(dividendData_1);
+      await dividendFacet.connect(signer_C).setDividend(dividendData);
 
       //  transfer
       const canRedeem = await erc1410Facet
@@ -1458,8 +1460,8 @@ describe("ERC1410 Tests", () => {
       expect(partitionsOf_E[0]).to.equal(_PARTITION_ID_1);
       expect(balanceOf_E).to.equal(balanceOf_E_Original - amount);
       expect(balanceOf_E_Partition_1).to.equal(balanceOf_E);
-      let dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      let dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      let dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      let dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(0);
       expect(dividend.snapshotId).to.equal(0);
       expect(totalSupply).to.be.equal(balanceOf_C_Original + balanceOf_E_Original - 2 * amount);
@@ -1474,8 +1476,8 @@ describe("ERC1410 Tests", () => {
         .withArgs(1, ethers.toBeHex(1, 32));
 
       // check that scheduled snapshots was triggered
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(1);
       expect(dividend.snapshotId).to.equal(0);
 
@@ -1492,8 +1494,8 @@ describe("ERC1410 Tests", () => {
         .withArgs(2, ethers.toBeHex(2, 32));
 
       // check that scheduled snapshots was triggered
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(1);
       expect(dividend.snapshotId).to.equal(2);
     });
@@ -1671,8 +1673,8 @@ describe("ERC1410 Tests", () => {
         amount: 1,
         amountDecimals: 0,
       };
-      await equityFacet.connect(signer_C).setDividend(dividendData_1);
-      await equityFacet.connect(signer_C).setDividend(dividendData);
+      await dividendFacet.connect(signer_C).setDividend(dividendData_1);
+      await dividendFacet.connect(signer_C).setDividend(dividendData);
 
       // controller transfer
       await expect(
@@ -1707,8 +1709,8 @@ describe("ERC1410 Tests", () => {
       const balanceOf_E_Partition_1 = await erc1410Facet.balanceOfByPartition(_PARTITION_ID_1, signer_E.address);
       expect(balanceOf_E).to.equal(balanceOf_E_Original + amount);
       expect(balanceOf_E_Partition_1).to.equal(balanceOf_E);
-      let dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      let dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      let dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      let dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(0);
       expect(dividend.snapshotId).to.equal(0);
 
@@ -1732,8 +1734,8 @@ describe("ERC1410 Tests", () => {
         .withArgs(1, ethers.toBeHex(1, 32));
 
       // check that scheduled snapshots was triggered
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(1);
       expect(dividend.snapshotId).to.equal(0);
 
@@ -1750,8 +1752,8 @@ describe("ERC1410 Tests", () => {
         .withArgs(2, ethers.toBeHex(2, 32));
 
       // check that scheduled snapshots was triggered
-      dividend_1 = (await equityFacet.getDividend(1)).registeredDividend_;
-      dividend = (await equityFacet.getDividend(2)).registeredDividend_;
+      dividend_1 = (await dividendFacet.getDividend(1)).registeredDividend_;
+      dividend = (await dividendFacet.getDividend(2)).registeredDividend_;
       expect(dividend_1.snapshotId).to.equal(1);
       expect(dividend.snapshotId).to.equal(2);
     });
