+++ packages/ats/contracts/test/contracts/integration/layer_1/adjustBalances/adjustBalances.test.ts
@@ -15,6 +15,7 @@ import {
   SsiManagement,
   TimeTravelFacet as TimeTravel,
   type MigrationFacetTest,
+  DividendFacet,
 } from "@contract-types";
 import { grantRoleAndPauseToken } from "@test";
 import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
@@ -43,6 +44,7 @@ describe("Adjust Balances Tests", () => {
   let timeTravelFacet: TimeTravel;
   let kycFacet: Kyc;
   let ssiManagementFacet: SsiManagement;
+  let dividendFacet: DividendFacet;
 
   async function deploySecurityFixtureMultiPartition() {
     const base = await deployEquityTokenFixture({
@@ -87,6 +89,7 @@ describe("Adjust Balances Tests", () => {
 
     kycFacet = await ethers.getContractAt("Kyc", diamond.target);
     ssiManagementFacet = await ethers.getContractAt("SsiManagement", diamond.target);
+    dividendFacet = await ethers.getContractAt("DividendFacet", diamond.target);
   }
 
   beforeEach(async () => {
@@ -154,7 +157,7 @@ describe("Adjust Balances Tests", () => {
       amountDecimals: dividendAmountDecimalsPerEquity,
     };
 
-    await equityFacet.connect(signer_A).setDividend(dividendData_1);
+    await dividendFacet.connect(signer_A).setDividend(dividendData_1);
 
     const balanceAdjustmentExecutionDateInSeconds_1 = dateToUnixTimestamp(`2030-01-01T00:00:07Z`);
