+++ packages/ats/contracts/test/contracts/integration/layer_1/scheduledTasks/scheduledTasks/scheduledTasks.test.ts
@@ -13,6 +13,7 @@ import {
   type IERC1410,
   Kyc,
   SsiManagement,
+  DividendFacet,
 } from "@contract-types";
 import { ZERO, EMPTY_STRING, dateToUnixTimestamp, ATS_ROLES, ATS_TASK } from "@scripts";
 import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
@@ -37,6 +38,7 @@ describe("Scheduled Tasks Tests", () => {
   let timeTravelFacet: TimeTravelFacet;
   let kycFacet: Kyc;
   let ssiManagementFacet: SsiManagement;
+  let dividendFacet: DividendFacet;
 
   async function deploySecurityFixtureSinglePartition() {
     const base = await deployEquityTokenFixture();
@@ -69,6 +71,7 @@ describe("Scheduled Tasks Tests", () => {
   async function setFacets(diamond: ResolverProxy) {
     accessControlFacet = await ethers.getContractAt("AccessControlFacet", diamond.target, signer_A);
     equityFacet = await ethers.getContractAt("EquityUSA", diamond.target, signer_A);
+    dividendFacet = await ethers.getContractAt("DividendFacet", diamond.target, signer_A);
     scheduledTasksFacet = await ethers.getContractAt("ScheduledCrossOrderedTasks", diamond.target, signer_A);
 
     pauseFacet = await ethers.getContractAt("Pause", diamond.target, signer_A);
@@ -127,8 +130,8 @@ describe("Scheduled Tasks Tests", () => {
       amount: dividendsAmountPerEquity,
       amountDecimals: dividendsAmountDecimalsPerEquity,
     };
-    await equityFacet.connect(signer_C).setDividend(dividendData_2);
-    await equityFacet.connect(signer_C).setDividend(dividendData_1);
+    await dividendFacet.connect(signer_C).setDividend(dividendData_2);
+    await dividendFacet.connect(signer_C).setDividend(dividendData_1);
 
     const balanceAdjustmentExecutionDateInSeconds_1 = dateToUnixTimestamp("2030-01-01T00:00:16Z");
     const balanceAdjustmentExecutionDateInSeconds_2 = dateToUnixTimestamp("2030-01-01T00:00:31Z");
@@ -171,8 +174,8 @@ describe("Scheduled Tasks Tests", () => {
     await timeTravelFacet.changeSystemTimestamp(balanceAdjustmentExecutionDateInSeconds_1 + 1);
 
     // Checking dividends For before triggering from the queue
-    const BalanceOf_A_Dividend_1 = await equityFacet.getDividendFor(2, signer_A.address);
-    let BalanceOf_A_Dividend_2 = await equityFacet.getDividendFor(1, signer_A.address);
+    const BalanceOf_A_Dividend_1 = await dividendFacet.getDividendFor(2, signer_A.address);
+    let BalanceOf_A_Dividend_2 = await dividendFacet.getDividendFor(1, signer_A.address);
 
     expect(BalanceOf_A_Dividend_1.tokenBalance).to.equal(INITIAL_AMOUNT);
     expect(BalanceOf_A_Dividend_2.tokenBalance).to.equal(0);
@@ -195,7 +198,7 @@ describe("Scheduled Tasks Tests", () => {
     // AFTER SECOND SCHEDULED SNAPSHOTS ------------------------------------------------------------------
     await timeTravelFacet.changeSystemTimestamp(balanceAdjustmentExecutionDateInSeconds_2 + 1);
     // Checking dividends For before triggering from the queue
-    BalanceOf_A_Dividend_2 = await equityFacet.getDividendFor(1, signer_A.address);
+    BalanceOf_A_Dividend_2 = await dividendFacet.getDividendFor(1, signer_A.address);
 
     expect(BalanceOf_A_Dividend_2.tokenBalance).to.equal(INITIAL_AMOUNT * balanceAdjustmentsFactor_1);
     expect(BalanceOf_A_Dividend_2.decimals).to.equal(DECIMALS_INIT + balanceAdjustmentsDecimals_1);
